+++ tools/automation/cli_linter/__init__.py
@@ -1,0 +1,96 @@
+# --------------------------------------------------------------------------------------------
+# Copyright (c) Microsoft Corporation. All rights reserved.
+# Licensed under the MIT License. See License.txt in the project root for license information.
+# --------------------------------------------------------------------------------------------
+
+
+import sys
+import yaml
+import os
+import inspect
+from knack.help_files import helps
+from importlib import import_module
+from pkgutil import iter_modules
+from .linter import LinterManager, _get_command_groups
+
+
+def define_arguments(parser):
+    parser.add_argument('--ci', action='store_true', help='Run in CI mode')
+    parser.add_argument('--params', dest='rule_types_to_run', action='append_const', const='params',
+                        help='Run linter on parameters.')
+    parser.add_argument('--commands', dest='rule_types_to_run', action='append_const', const='commands',
+                        help='Run linter on commands.')
+    parser.add_argument('--command-groups', dest='rule_types_to_run', action='append_const', const='command_groups',
+                        help='Run linter on command groups.')
+    parser.add_argument('--help-file-entries', dest='rule_types_to_run', action='append_const', const='help_entries',
+                        help='Run linter on help-file-entries.')
+    parser.add_argument('--modules', dest='modules', nargs='+', help='The modules on which the linter should run.')
+
+
+def main(args):
+    from azure.cli.core import get_default_cli
+    from azure.cli.core.file_util import get_all_help, create_invoker_and_load_cmds_and_args
+
+    print('Initializing linter with command table and help files...')
+    # setup CLI to enable command loader
+    az_cli = get_default_cli()
+
+    # load commands, args, and help
+    create_invoker_and_load_cmds_and_args(az_cli)
+    loaded_help = get_all_help(az_cli)
+    command_table = az_cli.invocation.commands_loader.command_table
+
+    # format loaded help
+    loaded_help = {data.command: data for data in loaded_help if data.command}
+
+    # load yaml help
+    help_file_entries = {}
+    for entry_name, help_yaml in helps.items():
+        help_entry = yaml.load(help_yaml)
+        help_file_entries[entry_name] = help_entry
+
+    if not args.rule_types_to_run:
+        args.rule_types_to_run = ['params', 'commands', 'command_groups', 'help_entries']
+
+    # find rule exclusions and pass to linter manager
+    from ..utilities.path import get_command_modules_paths
+    exclusions = {}
+    command_modules_paths = get_command_modules_paths()
+    for _, path in get_command_modules_paths():
+        exclusion_path = os.path.join(path, 'linter_exclusions.yml')
+        if os.path.isfile(exclusion_path):
+            mod_exclusions = yaml.load(open(exclusion_path))
+            exclusions.update(mod_exclusions)
+
+    # only run linter on modules specified
+    if args.modules:
+        allowed_module_paths = tuple(path for mod, path in command_modules_paths if mod in args.modules)
+        for command_name, command in list(command_table.items()):
+            # brute force way to remove all traces from excluded modules
+            loader_cls = command.loader.__class__
+            loader_file_path = inspect.getfile(loader_cls)
+            if not loader_file_path.startswith(allowed_module_paths):
+                del command_table[command_name]
+                help_file_entries.pop(command_name, None)
+                for group_name in _get_command_groups(command_name):
+                    help_file_entries.pop(group_name, None)
+
+
+    # Instantiate and run Linter
+    linter_manager = LinterManager(command_table=command_table,
+                                   help_file_entries=help_file_entries,
+                                   loaded_help=loaded_help,
+                                   exclusions=exclusions)
+    exit_code = linter_manager.run(run_params='params' in args.rule_types_to_run,
+                                   run_commands='commands' in args.rule_types_to_run,
+                                   run_command_groups='command_groups' in args.rule_types_to_run,
+                                   run_help_files_entries='help_entries' in args.rule_types_to_run,
+                                   ci=args.ci)
+
+    sys.exit(exit_code)
+
+
+def init_args(root):
+    parser = root.add_parser('cli-lint', help="Verify the contents of the command table and help.")
+    define_arguments(parser)
+    parser.set_defaults(func=main)